soluc

+++ database/migrations/2025_01_06_195453_add_empresaid_to_usuarios.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->foreignId('empresa_id')->nullable()->constrained('empresa')->onDelete('cascade');
+            if (!Schema::hasColumn('users', 'empresa_id')) {
+                $table->foreignId('empresa_id')->nullable()->constrained('empresa')->onDelete('cascade');
+            }
         });
     }
